requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
aiosqlite>=0.19.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0
email-validator>=2.1.0

# Telegram Integration
python-telegram-bot>=21.0
httpx>=0.26.0

# Rate Limiting & Security
slowapi>=0.1.9

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Media Processing (ffmpeg required separately)
# NOTE: Install FFmpeg on your system separately
